+++ src/kplslda.jl
@@ -1,0 +1,101 @@
+"""
+    kplslda(; kwargs...)
+    kplslda(X, y; kwargs...)
+    kplslda(X, y, weights::Weight; kwargs...)
+KPLS-LDA.
+* `X` : X-data (n, p).
+* `y` : Univariate class membership (n).
+* `weights` : Weights (n) of the observations. 
+    Must be of type `Weight` (see e.g. function `mweight`).
+Keyword arguments:
+* `nlv` : Nb. latent variables (LVs) to compute.
+    Must be >= 1
+* `kern` : Type of kernel used to compute the Gram matrices.
+    Possible values are: `:krbf`, `:kpol`. See respective 
+    functions `krbf` and `kpol` for their keyword arguments.
+* `prior` : Type of prior probabilities for class 
+    membership. Possible values are: `:unif` (uniform), 
+    `:prop` (proportional), or a vector (of length equal to 
+    the number of classes) giving the prior weight for each class 
+    (the vector must be sorted in the same order as `mlev(x)`).
+* `scal` : Boolean. If `true`, each column of `X` 
+    is scaled by its uncorrected standard deviation.
+
+Same as function `plslda` (PLS-LDA) except that 
+a kernel PLSR (function `kplsr`), instead of a 
+PLSR (function `plskern`), is run on the Y-dummy table. 
+
+## Examples
+```julia
+using JchemoData, JLD2
+path_jdat = dirname(dirname(pathof(JchemoData)))
+db = joinpath(path_jdat, "data/forages2.jld2")
+@load db dat
+pnames(dat)
+X = dat.X
+Y = dat.Y
+n = nro(X) 
+s = Bool.(Y.test)
+Xtrain = rmrow(X, s)
+ytrain = rmrow(Y.typ, s)
+Xtest = X[s, :]
+ytest = Y.typ[s]
+ntrain = nro(Xtrain)
+ntest = nro(Xtest)
+(ntot = n, ntrain, ntest)
+tab(ytrain)
+tab(ytest)
+
+nlv = 15
+gamma = .1
+mod = kplslda(; nlv, gamma) 
+#mod = kplslda(; nlv, gamma, prior = :prop) 
+#mod = kplsqda(; nlv, gamma) 
+fit!(mod, Xtrain, ytrain)
+pnames(mod)
+pnames(mod.fm)
+fm = mod.fm ;
+fm.lev
+fm.ni
+
+fmpls = fm.fm.fmpls ;
+@head fmpls.T
+@head transf(mod, Xtrain)
+@head transf(mod, Xtest)
+@head transf(mod, Xtest; nlv = 3)
+
+coef(fmpls)
+
+res = predict(mod, Xtest) ;
+pnames(res)
+@head res.posterior
+@head res.pred
+errp(res.pred, ytest)
+conf(res.pred, ytest).cnt
+
+predict(mod, Xtest; nlv = 1:2).pred
+```
+""" 
+function kplslda(X, y; kwargs...)
+    par = recovkwargs(Par, kwargs)
+    Q = eltype(X[1, 1])
+    weights = mweightcla(Q, y; prior = par.prior)
+    kplslda(X, y, weights; kwargs...)
+end
+
+function kplslda(X, y, weights::Weight; kwargs...)
+    par = recovkwargs(Par, kwargs)
+    @assert par.nlv >= 1 "Argument 'nlv' must be in >= 1"   
+    res = dummy(y)
+    ni = tab(y).vals
+    fmpls = kplsr(X, res.Y, weights; kwargs...)
+    fmda = list(Lda, par.nlv)
+    @inbounds for i = 1:par.nlv
+        fmda[i] = lda(fmpls.T[:, 1:i], y, weights; kwargs...)
+    end
+    fm = (fmpls = fmpls, fmda = fmda)
+    Plslda(fm, res.lev, ni)
+end
+
+
+