+++ docs/src/news.md
@@ -6,9 +6,8 @@
     - **aggsum** Compute sub-total sums by class for a categorical variable.
     - **findindex**: Replace a vector containg levels by the indexes of a set of levels.
     - **getknn**: Add of angular and correlation distances.
-    - **mbplskdeda**: Multiblock PLS-KDEDA.
-    - **mbplslda**: Multiblock PLS-LDA.
-    - **mbplsqda**: Multiblock PLS-QDA.
+    - **kplslda**, **kplsqda**, **kplskdeda**: Kernel PLS-LDA, PLS-QDA, PLS-KDEDA.
+    - **mbplslda**, **mbplsqda**, **mbplskdeda**: Multiblock PLS-LDA, PLS-QDA, PLS-KDEDA.
     - **merrp**: Mean intra-class classification error rate.
     - **mweightcla**: Compute observation weights for a categorical variable, 
         given specified sub-total weights for the classes.